requirements.txt
requests
beautifulsoup4
pytz
datetime
dataclasses-json
pandas
playwright
typing 
schedule 
python-telegram-bot
lxml
